Use semicolon entry separator in gcc LIBRARY_PATH environment variable (#2)
The installer script used to use a colon, which GCC docs say is the LIBRARY_PATH separator. This turns out to not be the case; LIBRARY_PATH uses semicolons as PATH does. Inserting colons caused the `LIBRARY_PATH ` environment variable to be interpreted incorrectly when appending on to an already existing `LIBRARY_PATH`.